+++ library/core/src/main/java/com/google/android/exoplayer2/ExoPlayerImpl.java
@@ -975,6 +975,12 @@ import java.util.concurrent.CopyOnWriteArraySet;
   }
 
   @Override
+  public MediaMetadata getMediaMetadata() {
+    // Unsupported operation.
+    return MediaMetadata.EMPTY;
+  }
+
+  @Override
   public Timeline getCurrentTimeline() {
     return playbackInfo.timeline;
   }